Support USB HCI
Issue: AXIA-1459

+++ btif/src/btif_util.c
@@ -94,8 +94,8 @@ char *bd2str(bt_bdaddr_t *bdaddr, bdstr_t *bdstr)
     char *addr = (char *) bdaddr->address;
 
     sprintf((char*)bdstr, "%02x:%02x:%02x:%02x:%02x:%02x",
-                       (int)addr[0],(int)addr[1],(int)addr[2],
-                       (int)addr[3],(int)addr[4],(int)addr[5]);
+                       (uint8_t)addr[0], (uint8_t)addr[1], (uint8_t)addr[2],
+                       (uint8_t)addr[3], (uint8_t)addr[4], (uint8_t)addr[5]);
     return (char *)bdstr;
 }
